requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
msal==1.25.0
requests==2.31.0
Office365-REST-Python-Client==2.5.3
gunicorn==21.2.0
openai==1.54.0
python-docx==1.1.2
pdfminer.six==20231228
tenacity==8.2.3